requirements.txt
python-telegram-bot
numpy==1.17.3
sklearn
pandas==1.3.1
pyyaml
ruamel.yaml==0.17.4
python-crontab